dialog handles string outcomes
If no values labels for the outcome variable, use FREQ to get the category values for the failure outcome value.
NOTE: still getting fatal errors sometime after the first survival command. So the branch is stuck here until I can sort it out. This is a new problem after building the dialog, but the error occurs when executing the survival command, not during the dialog