chart_options: removed dependency on Executor
ChartColorsFromOptions returns a message if any errors, else ''
Calling units must query the message and decide what to do. For now, all charts fail with an error message. This should be one of the first lines in a chart unit, so that no objects are created.
This could be handled by an exception, but the msg seems to be simpler.
For this to work, graphcommandexecutor must query a nil graph command result. Otherwise the following commands will fail, since nothing was returned.