Jamie/generic graph options (#53)
* restore missing paths

* Add generic chart options for colour and axis type

replace graph_utils with chart_options unit for generic options
added code to ast, including passing vector to optionlist.hasoption
revised color setting for all graph commands

* Update epidataanalysis.fpc

* chart_options: removed dependency on Executor

ChartColorsFromOptions returns a message if any errors, else ''
Calling units must query the message and decide what to do. For now, all charts fail with an error message. This should be one of the first lines in a chart unit, so that no objects are created.
This could be handled by an exception, but the msg seems to be simpler.
For this to work, graphcommandexecutor must query a nil graph command result. Otherwise the following commands will fail, since nothing was returned.

* add epi_osxlocale to LPR file (Darwin only)

epi_osxlocale should be in Core/Utils; until that is implemented, a duplicate is in analysis src/utils