Address spacing in chart options dialog; replace binary MinMax with defined types
Because MinMax dialog text / date fields can be hidden, we cannot reliably align labels to the entry boxes. To allow for linux font issue, I've increased spacing of the labels for linux only. The problem remains, that this might not work for all fonts. There is probably a way to get the height of an entry field, buy my reading suggests this is also not reliable. We could increase the spacing for ALL OS to 20 from 10.