Simplified Graph Save dialog
User MUST specify an extension (jpg, jpeg, svg or png). The drop down for file type has not worked properly, because the check on an existing file is ignored. The dialog only checked for the text provided. If this didn't have an extension, then an existing file would be overwritten.
Removes the exception for file type, which was overkill. A simple dialog is sufficient here.