Add development commands to grammar
This allows development of a new command without having to compile grammar. Will be useful once grammar can be compile automatically in gitHub.

cdev is used to execute the new command

in ast unit, cdev must point to the development command.
When ready for merge request, cdev must point to stNone and grammar must include the new command