Jamie/implement pareto (#60)
* Simplified Graph Save dialog

User MUST specify an extension (jpg, jpeg, svg or png). The drop down for file type has not worked properly, because the check on an existing file is ignored. The dialog only checked for the text provided. If this didn't have an extension, then an existing file would be overwritten.
Removes the exception for file type, which was overkill. A simple dialog is sufficient here.

* No chart created if no data

Must create the graph command result, but don't create or add charts if no data

* Implement Pareto

1. Pareto command; stratified produces multiple charts in one sheet
2. Implement multiple charts in one sheet
3. Trap exception caused by pareto - persistent axis
TODO:
1. Pareto dialog
2. Ignore DrawData Leak exception
3. How to save multiple charts from pareto command?

* Update grammar

Restore graph commands missing in this branch
Allow short commands for scatter, epicurve, histogram, barchart, pareto

* implement pareto dialog

* Update epidataanalysis.fpc

* Don't show tabs unless more than one chart

* Implement short commands in grammar

* Update graphform.impl.pas

* grammar rebuilt using command line tool

For some reason, the .inc file changes, but seems to work OK

* Implement chart font from set options

Note: Does not apply font to axis ticks (yet?)
Changes to charteaxesconfiguration and charttitles are first attempt to define right axis as a standard, since right axis will may be useful for charts other than pareto. However, none of that code is actually used, or effective, since Y2Axis is not connected to chart.axislist

* Destroy axis transformations before destroying chart

* Minor title corrections for KM plot and clipboard output

* Updated branch to current Laz and FPC