Fix output font not applied to html output
On the Mac, set "output font name" has never changed the font on screen or in a saved file.
Now, the standard embedded CSS has been changed from a constant to a constant string, so that it can be modified.
Changes work for Darwin build; I assume they will for Windows and Linux, but not verified.

lpi and lpr files not include Hi DPI awareness