Fix - save chart via !e:= causes crash
Problem: crash on freeing the chart pair when the chart was saved via command line
Solution: by trial and error, the problem seems to be trying to free the chart twice (1. when freeing the SaveAction object and 2. when freeing the ChartCommand object
Side effect: Chart transformations are not freed, causing a Draw Data Leak. Can't seem to eliminate this. A year ago, it seemed to go away if we free the axis  transformations first, but now it is probably because of some object left in saving the chart.