Implement save of multiple charts
All of these changes are related to the need to provide for multiple charts in SaveGraphAction, so that !export option and the save dialog for the graphs have the stratum value available to customize the chart file names. We could just name the files xxx-1.jpg, xxx-2.jpg, etc, but this has less meaning than xxx-StratumA.jpg, xxx-StratumB.jpg, etc.

NB: the chart command MUST provide the stratum values, as the default stratum value is ''.

With this commit, no commands yet do this.