Jamie/pareto save multiple charts (#82)
* Add stratum value to chart titles

Stratum value will be used to create the save file name for a command that produces multiple (tabbed) charts.

* Remove unnecessary var in chartpair

var i was left over from an attempt to destroy the axis transformations; this isn't the place to do that.

* Refactor savegraphaction.pas

1. Cannot call SaveGraphExecute directly from GraphCommandExecutor. So it now invokes SaveGraphs, which can be called from elsewhere. SaveGraphs must return true is the save was OK.
2. MacOS requires InitFonts to be able to get fonts for .SVG files.
3. Added destructor

* Implement save of multiple charts

All of these changes are related to the need to provide for multiple charts in SaveGraphAction, so that !export option and the save dialog for the graphs have the stratum value available to customize the chart file names. We could just name the files xxx-1.jpg, xxx-2.jpg, etc, but this has less meaning than xxx-StratumA.jpg, xxx-StratumB.jpg, etc.

NB: the chart command MUST provide the stratum values, as the default stratum value is ''.

With this commit, no commands yet do this.

* Pareto - enable saving of multiple charts

Implements previous multiple chart output in the pareto command.

* Address  comments

pareto: charts are labelled with '', 1, 2, 3...
graphcommandexecutor: options applied only if they exist
savegraphaction: drop validation of filename. The OS will give an exception if the file name is illegal and this is reported to the user.