Summary of model deviances for logistic regression
With !sum option, regress (logistic only) will save some model outputs into a secondary dataset (_lrSummary), creating this if it does not already exist.

The assumption is that every model run on the dataset will have the same dependent variable. If this is not a valid assumption, then we will need to let the user specify the name of the dataset used to track results.